# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "กรองตามหมวดหมู่",
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          },
          {
            "name": "isPreorder",
            "required": false,
            "in": "query",
            "description": "กรองเฉพาะพรีออเดอร์ (true) หรือพร้อมส่ง (false)",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "ค้นจากชื่อหรือรายละเอียด",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "หน้าที่ต้องการ เริ่มที่ 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "จำนวนต่อหน้า 1–100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductModel"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "รายการสินค้า (แบ่งหน้า)",
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProductModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เพิ่มสินค้า (เจ้าหน้าที่)",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProductModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "สินค้าหนึ่งรายการพร้อมตัวเลือก",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProductModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "แก้ไขสินค้าบางฟิลด์ (เจ้าหน้าที่)",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ลบสินค้า (ผู้ดูแล) — ลบไม่ได้ถ้าเคยถูกสั่งซื้อแล้ว",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/product-categories": {
      "get": {
        "operationId": "ProductCategoriesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "หน้าที่ต้องการ เริ่มที่ 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "จำนวนต่อหน้า 1–100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductCategoryModel"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "หมวดหมู่สินค้าพร้อมจำนวนสินค้าในแต่ละหมวด",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/product-variants": {
      "get": {
        "operationId": "ProductVariantsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "หน้าที่ต้องการ เริ่มที่ 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "จำนวนต่อหน้า 1–100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "description": "กรองตามสินค้า",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "inStock",
            "required": false,
            "in": "query",
            "description": "เฉพาะตัวเลือกที่ยังมีของ",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/VariantWithProductModel"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "รายการตัวเลือกสินค้า (ใช้หน้าจัดการสต็อก)",
        "tags": [
          "product-variants"
        ]
      },
      "post": {
        "operationId": "ProductVariantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVariantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/VariantWithProductModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เพิ่มตัวเลือกให้สินค้า (เจ้าหน้าที่)",
        "tags": [
          "product-variants"
        ]
      }
    },
    "/api/v1/product-variants/{id}": {
      "get": {
        "operationId": "ProductVariantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/VariantWithProductModel"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "product-variants"
        ]
      },
      "patch": {
        "operationId": "ProductVariantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVariantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/VariantWithProductModel"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "product-variants"
        ]
      },
      "delete": {
        "operationId": "ProductVariantsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedModel"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "product-variants"
        ]
      }
    },
    "/api/v1/product-variants/{id}/stock": {
      "patch": {
        "operationId": "ProductVariantsController_updateStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/VariantWithProductModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ปรับสต็อก (เพิ่ม/ลด หรือกำหนดค่า)",
        "tags": [
          "product-variants"
        ]
      }
    },
...
  },
  "info": {
    "title": "CSMJU Shop API",
    "description": "ร้านค้าออนไลน์สาขาวิทยาการคอมพิวเตอร์ ม.แม่โจ้ — ทุก endpoint ต้องมี Core Hub token (Authorization: Bearer หรือคุกกี้ core_hub_access_token) ยกเว้น GET /api/health · ทุก response ห่อด้วย { success, data, meta? } · เงินทุกช่องเป็นจำนวนเต็มหน่วยสตางค์",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PageMetaModel": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
      "ProductVariantModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "variantName": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "ราคาต่อชิ้น หน่วยสตางค์"
          },
          "stockQuantity": {
            "type": "number",
            "description": "ติดลบได้เฉพาะสินค้าพรีออเดอร์ (= ยอดที่ต้องสั่งผลิตเพิ่ม)"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "productId",
          "variantName",
          "price",
          "stockQuantity",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductModel": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "preorderEndDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "estimatedDelivery": {
            "type": "string",
            "nullable": true,
            "description": "YYYY-MM-DD",
            "format": "date"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantModel"
            }
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "isPreorder": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "description",
          "imageUrl",
          "preorderEndDate",
          "estimatedDelivery",
          "variants",
          "id",
          "name",
          "category",
          "isPreorder",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateProductVariantInlineDto": {
        "type": "object",
        "properties": {
          "variantName": {
            "type": "string",
            "description": "ชื่อตัวเลือก เช่น \"M\" หรือ \"ดำ / L\"",
            "minLength": 1,
            "maxLength": 50
          },
          "price": {
            "type": "number",
            "description": "ราคาต่อชิ้น หน่วยสตางค์ (เช่น 25000 = 250 บาท)",
            "minimum": 0
          },
          "stockQuantity": {
            "type": "number",
            "description": "จำนวนคงเหลือเริ่มต้น",
            "minimum": 0
          }
        },
        "required": [
          "variantName",
          "price"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "ชื่อสินค้า",
            "minLength": 1,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "description": "รายละเอียดสินค้า"
          },
          "category": {
            "type": "string",
            "description": "หมวดหมู่ เช่น \"เสื้อ\" \"ของที่ระลึก\"",
            "minLength": 1,
            "maxLength": 50
          },
          "imageUrl": {
            "type": "string",
            "description": "URL รูปที่ได้จาก POST /api/v1/product-images (ส่ง \"\" เพื่อเอารูปออก)",
            "minLength": 0,
            "maxLength": 255
          },
          "isPreorder": {
            "type": "boolean",
            "description": "เป็นสินค้าพรีออเดอร์หรือไม่"
          },
          "preorderEndDate": {
            "type": "string",
            "description": "วันปิดรับพรีออเดอร์ (ISO 8601) — บังคับเมื่อ isPreorder = true · ส่ง \"\" เพื่อล้างค่า"
          },
          "estimatedDelivery": {
            "type": "string",
            "description": "วันที่คาดว่าจะได้รับสินค้า (YYYY-MM-DD) · ส่ง \"\" เพื่อล้างค่า",
            "pattern": "^(\\d{4}-\\d{2}-\\d{2})?$"
          },
          "variants": {
            "description": "ตัวเลือกสินค้าเริ่มต้น",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateProductVariantInlineDto"
            }
          }
        },
        "required": [
          "name",
          "category"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "ชื่อสินค้า",
            "minLength": 1,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "description": "รายละเอียดสินค้า"
          },
          "category": {
            "type": "string",
            "description": "หมวดหมู่ เช่น \"เสื้อ\" \"ของที่ระลึก\"",
            "minLength": 1,
            "maxLength": 50
          },
          "imageUrl": {
            "type": "string",
            "description": "URL รูปที่ได้จาก POST /api/v1/product-images (ส่ง \"\" เพื่อเอารูปออก)",
            "minLength": 0,
            "maxLength": 255
          },
          "isPreorder": {
            "type": "boolean",
            "description": "เป็นสินค้าพรีออเดอร์หรือไม่"
          },
          "preorderEndDate": {
            "type": "string",
            "description": "วันปิดรับพรีออเดอร์ (ISO 8601) — บังคับเมื่อ isPreorder = true · ส่ง \"\" เพื่อล้างค่า"
          },
          "estimatedDelivery": {
            "type": "string",
            "description": "วันที่คาดว่าจะได้รับสินค้า (YYYY-MM-DD) · ส่ง \"\" เพื่อล้างค่า",
            "pattern": "^(\\d{4}-\\d{2}-\\d{2})?$"
          }
        }
      },
      "DeletedModel": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "id"
        ]
      },
      "ProductCategoryModel": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string"
          },
          "productCount": {
            "type": "number"
          }
        },
        "required": [
          "category",
          "productCount"
        ]
      },
      "ProductSummaryModel": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "isPreorder": {
            "type": "boolean"
          }
        },
        "required": [
          "imageUrl",
          "id",
          "name",
          "category",
          "isPreorder"
        ]
      },
      "VariantWithProductModel": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/ProductSummaryModel"
          },
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "variantName": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "ราคาต่อชิ้น หน่วยสตางค์"
          },
          "stockQuantity": {
            "type": "number",
            "description": "ติดลบได้เฉพาะสินค้าพรีออเดอร์ (= ยอดที่ต้องสั่งผลิตเพิ่ม)"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "product",
          "id",
          "productId",
          "variantName",
          "price",
          "stockQuantity",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateVariantDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "id ของสินค้า (UUID)",
            "format": "uuid"
          },
          "variantName": {
            "type": "string",
            "description": "ชื่อตัวเลือก เช่น \"M\"",
            "minLength": 1,
            "maxLength": 50
          },
          "price": {
            "type": "number",
            "description": "ราคาต่อชิ้น หน่วยสตางค์",
            "minimum": 0
          },
          "stockQuantity": {
            "type": "number",
            "description": "จำนวนคงเหลือเริ่มต้น",
            "minimum": 0
          }
        },
        "required": [
          "productId",
          "variantName",
          "price"
        ]
      },
      "UpdateVariantDto": {
        "type": "object",
        "properties": {
          "variantName": {
            "type": "string",
            "description": "ชื่อตัวเลือก เช่น \"M\"",
            "minLength": 1,
            "maxLength": 50
          },
          "price": {
            "type": "number",
            "description": "ราคาต่อชิ้น หน่วยสตางค์",
            "minimum": 0
          },
          "stockQuantity": {
            "type": "number",
            "description": "จำนวนคงเหลือเริ่มต้น",
            "minimum": 0
          }
        }
      },
      "UpdateStockDto": {
        "type": "object",
        "properties": {
          "adjust": {
            "type": "number",
            "description": "จำนวนที่เพิ่ม (บวก) หรือลด (ลบ)"
          },
          "set": {
            "type": "number",
            "description": "กำหนดจำนวนคงเหลือใหม่",
            "minimum": 0
          }
        }
      },
...
    }
  }
}